Green

+++ tests/cucumber/project.rs
@@ -13,11 +13,11 @@ impl Project {
     pub(crate) fn get_contribution_to(&self, filename: String) {
         let file_in_project = self.relative_path_to_project.join(filename);
         let mut file = File::options()
+            .create(true)
             .append(true)
             .open(file_in_project)
             .unwrap();
         writeln!(&mut file, "a").unwrap();
-
     }
 }
 
@@ -28,8 +28,10 @@ impl Project {
         if let Some(parent_dir) = file_in_project.parent() {
             create_dir_all(parent_dir).expect("Failed to create parent directory")
         }
-        File::create(file_in_project).unwrap();
-    }
+        if !file_in_project.exists() {
+            File::create(file_in_project).unwrap();
+        }
+        }
 }
 
 impl Project {
